Inline LoadDeckCounts inside DeckPicker
I had to disable the cancelling of this task in onPause, this is temporary, will be done in the next commits.
